restructure xml code
-move xml functions to another module
-beautify some code
